refactor(blob): lift documents/blob.rs to providers/blob.rs; ListOptions/ReadOptions/BlobEntry
Moves the documents connector's local-vs-object-store I/O module up to
providers/blob.rs so it can be shared with the upcoming obsidian source.
Adds ListOptions (recursive + follow_symlinks) and ReadOptions
(follow_symlinks) to control directory walking and file reads, and
BlobEntry to carry size/mtime alongside each listed location — all
needed by the vault scanner, which must never follow symlinks.
